requirements.txt
pandas
seaborn
streamlit
matplotlib
google-generativeai